requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
click>=8.0
Pillow>=9.0
face-recognition-models>=0.3.0
dlib-bin>=20.0.0; platform_system == "Windows"
setuptools<81
